Rename folder heidelberg_confession to heidelberg_catechism
